✨ *feat(bot): adicionar novas tarefas e otimizações*
Implementa novas funcionalidades e otimizações nas tarefas do bot, incluindo a adição de novos métodos para upload de arquivos e execução de tarefas.

Files Changed:
- *celery_app/tasks/__init__.py*: Atualiza a lista de exportação para incluir `_main`.
- *celery_app/tasks/_email.py*: Define o nome da tarefa para `send_email`.
- *celery_app/tasks/bot/_main.py*: Renomeia e registra a tarefa `run_task` como `run_bot`.
- *celery_app/tasks/bot/__init__.py*: Remove tarefas não utilizadas e registra a tarefa `print_message`.
- *celery_app/tasks/files.py*: Adiciona novas tarefas para upload e salvamento de arquivos.
